+++ scripts/test-helper.sh
@@ -35,6 +35,10 @@ ensure_submodule() {
   )
 }
 
+ensure_i686() {
+  export PKG_CONFIG_SYSROOT_DIR=/usr/lib/i386-linux-gnu/pkgconfig
+}
+
 # check_*_api <prefix> <features> <clippy-args>..
 # <prefix> = "api-", "applet-api-", or "board-api-"
 # <features> = "--features=" or "--features=wasm,std,"